Fix issues noted on meeting demo #LMR-510

+++ src/app/core/rest/project.service.ts
@@ -26,34 +26,63 @@ import {Project} from '../dto';
 import {FetchFailedError} from '../error/fetch-failed.error';
 import {NetworkError} from '../error/network.error';
 import {PermissionService} from './permission.service';
+import {LumeerError} from '../error/lumeer.error';
 
 @Injectable()
 export class ProjectService extends PermissionService {
 
   public getProjects(orgCode: string): Observable<Project[]> {
+    if (!this.hasOrganizationApiPrefix(orgCode)) {
+      throw new LumeerError('Organization not set');
+    }
+
     return this.httpClient.get<Project[]>(this.apiPrefix(orgCode)).pipe(
       catchError(ProjectService.catchGetProjectsError)
     );
   }
 
   public getProject(orgCode: string, projCode: string): Observable<Project> {
+    if (!this.hasFullApiPrefix(orgCode, projCode)) {
+      throw new LumeerError(`Workspace not set ${orgCode} ${projCode}`);
+    }
+
     return this.httpClient.get<Project>(this.apiPrefix(orgCode, projCode));
   }
 
   public deleteProject(orgCode: string, projCode: string): Observable<HttpResponse<any>> {
+    if (!this.hasFullApiPrefix(orgCode, projCode)) {
+      throw new LumeerError(`Workspace not set ${orgCode} ${projCode}`);
+    }
+
     return this.httpClient.delete(this.apiPrefix(orgCode, projCode), {observe: 'response', responseType: 'text'});
   }
 
   public createProject(orgCode: string, project: Project): Observable<Project> {
+    if (!this.hasOrganizationApiPrefix(orgCode)) {
+      throw new LumeerError('Organization not set');
+    }
+
     return this.httpClient.post(this.apiPrefix(orgCode), project, {observe: 'response', responseType: 'text'}).pipe(
       map(() => project) // TODO return fresh instance from the server instead
     );
   }
 
   public editProject(orgCode: string, projCode: string, project: Project): Observable<Project> {
+    if (!this.hasFullApiPrefix(orgCode, projCode)) {
+      throw new LumeerError(`Workspace not set ${orgCode} ${projCode}`);
+    }
+
     return this.httpClient.put(this.apiPrefix(orgCode, projCode), project, {observe: 'response', responseType: 'text'}).pipe(
       map(() => project) // TODO return fresh instance from the server instead
     );
+  }
+
+  private hasOrganizationApiPrefix(orgCode: string): boolean {
+    return !!orgCode;
+  }
+
+  private hasFullApiPrefix(orgCode: string, projCode: string): boolean {
+    return !!(orgCode && projCode);
   }
 
   private apiPrefix(orgCode: string, projCode?: string): string {
